:sparkles: feat(shell): Replace old dotfiles with new project
This commit replaces the old dotfiles with an updated project that
includes a new configuration for Zsh, Neovim, and Git. It also
introduces separate files for aliases, exports, and functions, and adds
an install script for easy setup.

The new dotfiles are designed to work on both macOS and Linux systems
and provide a cleaner, more organized structure for better maintainability.

Issue: Closes #1